Move season calculation utils to separate module
